backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
psycopg2-binary==2.9.7
Flask-Cors==4.0.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0 